login/requirements.txt
Django>=3.2,<4.0
psycopg2-binary
django-bootstrap-v5
djangorestframework
markdown
django-filter
django-cors-headers
djangorestframework-simplejwt
django-two-factor-auth